merge

+++ src/pages/devices.vue
@@ -94,8 +94,9 @@
                             <td>
                                 <ul class="ulactions">
                                     <li>
-                                        <router-link to="/modifyDevice">
-                                         <input type="button" class="btn-flat primary" value="修改" @click="modifyDevice($event)"/>
+                                    
+                                        <router-link :to='{name:"modifyDevice",params:{id:device.id}}'>
+                                         <input type="button" class="btn-flat primary" value="修改"/>
                                         </router-link>
                                     </li>
                                     <li>
@@ -138,9 +139,6 @@
 
 <script>
 /* eslint-disable */
-import modifyDevice from '@/pages/modifyDevice'
-
-/*let projectId = "5a922835-a587-4dad-b3b7-bb5005ef4c99";*/
 
 export default{
         data(){
@@ -176,31 +174,59 @@ export default{
                 var e = event || window.event;
                 //alert("B");
                 var target = e.target || e.srcElement;
-                if (target.parentNode.parentNode.parentNode.tagName.toLowerCase() == "td") {
-                    //alert("C");
-                    var rowIndex = target.parentNode.parentNode.parentNode.parentNode.rowIndex;
-                    //alert(rowIndex);
-                    var id = document.getElementById("table_value").rows[rowIndex].cells[0].innerHTML;
-                    //alert(id);
-                    var qs = require('qs');
-                    this.$axios.delete('devices/'+id,{
 
-                        //设置头
-                        headers:{
-                            'content-type':'application/x-www-form-urlencoded'
-                        },
-                        auth: {
-                            username: 'admin',
-                            password: 'admin'
-                        }
-                    }).then(res=>{
-                        //this.users = res.data.data
-                        //console.log(res);
-                        this.$router.replace({ path: '/devices'})
-                    }).catch(err=>{
-                        alert("删除失败！");
-                    })
+                var msg = "您确定删除吗？";
+                if (confirm(msg) == true) {
+                    
+                    if (target.parentNode.parentNode.parentNode.tagName.toLowerCase() == "td") {
+                        //alert("C");
+                        var rowIndex = target.parentNode.parentNode.parentNode.parentNode.rowIndex;
+                        //alert(rowIndex);
+                        var id = document.getElementById("table_value").rows[rowIndex].cells[0].innerHTML;
+                        //alert(id);
+                        var qs = require('qs');
+                        this.$axios.delete('devices/'+id,{
+
+                            //设置头
+                            headers:{
+                                'content-type':'application/x-www-form-urlencoded'
+                            },
+                            auth: {
+                                username: 'admin',
+                                password: 'admin'
+                            }
+                        }).then(res=>{
+
+                            //删除完再次查询
+                            var projectId = this.getCookie('projectId');
+                            var username = this.getCookie('username');
+                            var password = this.getCookie('password');
+                            this.$axios.get('project/'+projectId+'/device',{
+                                //设置头
+                                headers:{
+                                    'content-type':'application/x-www-form-urlencoded'
+                                },
+                                auth: {
+                                    username: username,
+                                    password: password
+                                }
+                            }).then(res=>{
+                                this.devices = res.data.data
+                            })
+                            .catch(err=>{
+                                console.log(err);
+                            })
+
+                        }).catch(err=>{
+                            alert("删除失败！");
+                        })
+                    }
+                    
+                } else {
+                    return false;
                 }
+
+                
 
             },
 
@@ -263,39 +289,7 @@ export default{
                 })
             }
 
-            /*,
-
-            modifyDevice: function (event){
-                alert("A");
-                var e = event || window.event;
-                //alert("B");
-                var target = e.target || e.srcElement;
-                if (target.parentNode.parentNode.parentNode.parentNode.tagName.toLowerCase() == "td") {
-                    //alert("C");
-                    var rowIndex = target.parentNode.parentNode.parentNode.parentNode.parentNode.rowIndex;
-                    //alert(rowIndex);
-                    var id = document.getElementById("table_value").rows[rowIndex].cells[0].innerHTML;
-                    alert(id);
-                    var qs = require('qs');
-                    this.$axios.get('devices/'+id,
-                      {
-                          //设置头
-                          headers:{
-                              'content-type':'application/x-www-form-urlencoded'
-                          },
-                          auth: {
-                              username: 'admin',
-                              password: 'admin'
-                          }
-                      }).then(res=>{
-                          console.log(res);
-                      })
-                      .catch(err=>{
-                          console.log(err);
-                      })
-                }
-
-            }*/
+            
         },
         computed: {  
             devicesA: function () {
